added tests for server, and established template for game tests
