+++ .gitignore
@@ -48,6 +48,9 @@ coverage.xml
 .hypothesis/
 .pytest_cache/
 
+#Mocking 
+mocking*
+
 # Translations
 *.mo
 *.pot
