+++ requirements.txt
@@ -1,12 +1,19 @@
+blinker==1.4
 certifi==2018.4.16
+chardet==3.0.4
 Flask==0.10.1
 Flask-DebugToolbar==0.10.0
 Flask-SQLAlchemy==2.0
+idna==2.6
 itsdangerous==0.24
 Jinja2==2.7.3
+linecache2==1.0.0
+MarkupSafe==1.1.0
 psycopg2==2.7.6
 requests==2.18.4
+six==1.11.0
 SQLAlchemy==1.0.3
+traceback2==1.4.0
+unittest2==1.1.0
 urllib3==1.22
 Werkzeug==0.10.4
-
