+++ test_game.py
@@ -1,0 +1,23 @@
+from game import Game
+import unittest
+from unittest.mock import patch
+
+
+class GameTests(unittest.TestCase):
+
+	def setUp(self,my_mock):
+		pass
+
+	def test_get_word(self):
+		pass
+	
+	def test_get_word_length(self):
+		pass
+	
+
+
+
+if __name__ == '__main__':  # pragma: no cover
+
+    import unittest
+    unittest.main()
